docs: cleaning up migration docs
