+++ docs/src/pages/_app.page.tsx
@@ -1,15 +1,12 @@
 import * as React from 'react';
-import { Storage, Amplify } from 'aws-amplify';
 import { useRouter } from 'next/router';
 import Script from 'next/script';
 
 import { ThemeProvider, ColorMode, defaultTheme } from '@aws-amplify/ui-react';
 
-import MyStorageProvider from '@/utils/storageMock';
 import { configure, trackPageVisit } from '@/utils/track';
 import { Header } from '@/components/Layout/Header';
 import { baseTheme } from '../theme';
-import { mockConfig } from '../mockConfig';
 
 import { Head } from './Head';
 
@@ -44,12 +41,6 @@ if (typeof window === 'undefined') {
   ✨ you can explore the Amplify UI theme object by typing \`theme\` in the console.
  `);
   window['theme'] = defaultTheme;
-  Amplify.configure(mockConfig);
-  Storage.addPluggable(new MyStorageProvider('fast', { delay: 10 }));
-  Storage.addPluggable(new MyStorageProvider('slow', { delay: 1000 }));
-  Storage.addPluggable(
-    new MyStorageProvider('error', { delay: 50, networkError: true })
-  );
 }
 
 function MyApp({ Component, pageProps }) {
